refactor: 노트 모아보기 관련 파일들을 액션 단위 slice 폴더로 분리
