+++ src/entities/note/index.ts
@@ -1,4 +1,2 @@
-// Note 엔티티 Public API
-export * from './model';
-// export * from './api';
-// export * from './ui';
+export { noteKeys } from './model/queryKeys';
+export type { NoteListResponse, NoteResponse } from './model/types';
