--- src/entities/note/model/index.ts
@@ -1,1 +1,0 @@
-export { type Note, type NoteListResponse } from './types';
