+++ src/entities/note/model/queryKeys.ts
@@ -1,0 +1,7 @@
+export const noteKeys = {
+  all: ['notes'] as const,
+  lists: () => [...noteKeys.all, 'list'] as const,
+  list: (studyGoalId: number) => [...noteKeys.lists(), studyGoalId] as const,
+  details: () => [...noteKeys.all, 'detail'] as const,
+  detail: (id: number) => [...noteKeys.details(), id] as const
+};
