--- src/features/note/api/index.ts
@@ -1,2 +1,0 @@
-export { useNotesByStudyGoalId, useNoteById } from './noteQueries';
-export { getNotesByStudyGoalId, getNoteById } from './noteApi';
