+++ src/features/note/get-note/index.ts
@@ -1,0 +1,1 @@
+export { NoteCard } from './ui/NoteCard';
