+++ src/features/note/get-note/ui/NoteCard.tsx
