--- src/features/note/index.ts
@@ -1,4 +1,0 @@
-// Note 피처 Public API
-export * from './model';
-// export * from './api';
-// export * from './ui';
