--- src/features/note/model/index.ts
@@ -1,2 +1,0 @@
-export { noteKeys } from './queryKeys';
-export { type Note, type NoteListResponse, type NoteResponse } from './types';
