--- src/features/note/model/types.ts
@@ -1,4 +1,0 @@
-
-
-
-
