--- src/features/note/ui/index.ts
@@ -1,2 +1,0 @@
-export { NoteCard } from './NoteCard';
-export { NoteEditor } from '../edit-note/ui/NoteEditor';
